// comp-graphs/src/types/Color.ts
enum Color {
    Blue = 'Blue',
    Crimson = 'Crimson',
    DarkCyan = 'DarkCyan',
    DarkGreen = 'DarkGreen',
    DarkOrange = 'DarkOrange',
    DodgerBlue = 'DarkBlue',
    Gold = 'Gold',
    Green = 'Green',
    LawnGreen = 'LawnGreen',
    LightSeaGreen = 'LightSeaGreen',
    MediumBlue = 'MediumBlue',
    Orange = 'Orange',
    Red = 'Red',
    Purple = 'Purple',
    SteelBlue = 'SteelBlue',
    Indigo = 'Indigo',
    FireBrick = 'FireBrick',
    DarkMagenta = 'DarkMagenta'
};

export default Color;